% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codesBR.R
\docType{package}
\name{codesBR-package}
\alias{codesBR}
\alias{codesBR-package}
\title{R functions to add Brazilian municipal Identifiers to a table}
\description{
The codesBR provides simple functions to add Brazilian municipal identifiers from IBGE, TSE, and SIAFI to an existing table, based on official data sources, in a tidy way.
}
\details{
To learn more about codesBR, check the package's documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meirelesff/genderBR}
  \item Report bugs at \url{https://github.com/meirelesff/genderBR/issues}
}

}
\author{
Fernando Meireles (UFMG/Brazil) [author]
}
