% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibge_from_tse.R
\name{ibge_from_tse}
\alias{ibge_from_tse}
\title{Add IBGE identifiers to a table}
\usage{
ibge_from_tse(data, tse_code)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tbl_df} object contaning a \code{character} variable with TSE municipal identifiers.}

\item{tse_code}{A (\code{character}) variable containing TSE municipal identifiers.}
}
\value{
The original table with IBGE municipality identifiers (as \code{character} variable). Multiple observations with the same identifier in the original data are preserved (using dplyr's left_join function).
}
\description{
The \code{ibge_from_tse} is a convinience function that takes a table and a variable with TSE municipal identifiers as inputs and adds a new variable to the original table containing the correspondent IBGE municipal identifiers.
}
\note{
The number of municipalities in Brazil increased markedly from 1980 to 2000. The data used in this package only covers the more recent composition of municipalities in the country.
}
\examples{
\dontrun{
# Using a table 'df' with the variable 'var_id' containing municipality identifiers
 df <- ibge_from_tse(df, var_id)
}

}
