% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_codes.R
\name{join_codes}
\alias{join_codes}
\title{Add IBGE, TSE, SIAFI, and other Brazilian municipal identifiers to a table}
\usage{
join_codes(data, code_var, by_code = c("ibge", "tse", "siafi"),
  states = FALSE, regions = FALSE, muni_names = FALSE, latlon = FALSE,
  ascii = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tbl_df} object contaning a \code{character} variable with IBGE or TSE municipality identifiers.}

\item{code_var}{A (\code{character}) variable containing the IBGE, TSE, or SIAFI identifiers.}

\item{by_code}{A (\code{character}) variable denoting the type of identifier passed to the function call. Possible values are: \code{'ibge'}, to use IBGE municipality identifiers; \code{'tse'}; and \code{'siafi'}.}

\item{states}{Should state information be added to the existing data? Defaults to \code{FALSE}. If set to \code{TRUE}, states' names, acronyims, and their respective IBGE identifiers are added to the original data.}

\item{regions}{Should mesorregion and microrregion information be added to the existing data? Defaults to \code{FALSE}. If set to \code{TRUE}, information on the mesorregion and microrregion to, acronyims, their respective IBGE identifiers are added to the original data.}

\item{muni_names}{Should municipalities' full names be added to the existing data? Defaults to \code{FALSE}.}

\item{latlon}{Should latitude and longitute information be added to the existing data? Defaults to \code{FALSE}.}

\item{ascii}{Convert results from 'latin1' to 'ascii'? Defaults to \code{FALSE}.}
}
\value{
The \code{join_codes} returns the original table with additional municipality identifiers (as \code{character} variables). Multiple observations with the same identifier in the original data are preserved (using dplyr's left_join function).
}
\description{
The \code{join_codes} takes a table and a variable with IBGE, TSE or SIAFI municipality identifiers as inputs and adds new variables to the original table, such as state, region, and municipal identifiers provided by IBGE, TSE, and SIAFI.
}
\note{
The number of municipalities in Brazil increased markedly from 1980 to 2000. The data used in this package covers the recent composition of municipalities in the country.
}
\section{Data sources}{
 The \code{codesBR} use data from official sources, such as IBGE (<https://concla.ibge.gov.br/classificacoes/por-tema/codigo-de-areas>), TSE (<http://www.tse.jus.br/arquivos/tse-lista-de-municipios-do-cadastro-da-justica-eleitoral/view>), Interlegis (<http://www.interlegis.leg.br/produtos_servicos/informacao/censo>), and SIAFI (<www.tesouro.fazenda.gov.br/>), to join Brazilian municipality codes to a previous table. The internal data contains 5,570 observations, covering all Brazilian municipalities (including Brazil's Fernando de Noronha e Brasilia).
}

\examples{
\dontrun{
# Using a table 'df' with the variable 'var_id' containing municipality identifiers
 df <- join_codes(df, var_id, "ibge")

# Get state identifiers
df <- join_codes(df, var_id, "ibge", states = TRUE)

# Get region identifiers
df <- join_codes(df, var_id, "ibge", regions = TRUE)

# Get the full name of the municipalities
df <- join_codes(df, var_id, "ibge", muni_names = TRUE)

#' # Get the lat and lon information
df <- join_codes(df, var_id, "ibge", latlon = TRUE)

# Get results in ASCII
df <- join_codes(df, var_id, "ibge", ascii = TRUE)
}

}
