% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse_from_ibge.R
\name{tse_from_ibge}
\alias{tse_from_ibge}
\title{Add TSE identifiers to a table}
\usage{
tse_from_ibge(data, ibge_code)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tbl_df} object contaning a \code{character} variable with IBGE municipal identifiers.}

\item{ibge_code}{A (\code{character}) variable containing IBGE municipal identifiers.}
}
\value{
The original table with TSE municipality identifiers (as \code{character} variable). Multiple observations with the same identifier in the original data are preserved (using dplyr's left_join function).
}
\description{
The \code{tse_from_ibge} is a convinience function that takes a table and a variable with IBGE municipal identifiers as inputs and adds a new variable to the original table containing the correspondent TSE municipal identifiers.
}
\note{
The number of municipalities in Brazil increased markedly from 1980 to 2000. The data used in this package only covers the more recent composition of municipalities in the country.
}
\examples{
\dontrun{
# Using a table 'df' with the variable 'var_id' containing municipality identifiers
 df <- tse_from_ibge(df, var_id)
}

}
